requirements.txt
# Framework web
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Werkzeug==2.3.7
gunicorn==21.2.0

# Seguridad y encriptación
cryptography==41.0.4
python-dotenv==1.0.0

# Utilidades adicionales
click==8.1.7
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3

# Base de datos (para producción)
psycopg2-binary==2.9.7

# Desarrollo y testing (opcional)
pytest==7.4.3
pytest-flask==1.3.0
